Make template dynamic based on the value of the key
It's possible to define additional versions of a template based on the
current value of the key. This allows for more dynamic templates based
on the previous user's input. For instance it can be used to indicate
that a server would be uninstalled if "skip" is chosen during a
reconfiguration.

Also:
- Change assert output to make diff easier to reader
- Keep paragraphs in the description: separate them with 2 newlines
  instead of only 1.